+++ example/wallpaper.dart
@@ -58,18 +58,10 @@ void printBackgroundColor() {
 void main() {
   final hr = CoInitializeEx(
       nullptr, COINIT_APARTMENTTHREADED | COINIT_DISABLE_OLE1DDE);
-
-  if (FAILED(hr)) {
-    throw WindowsException(hr);
-  }
+  if (FAILED(hr)) throw WindowsException(hr);
 
   wallpaper = DesktopWallpaper.createInstance();
 
-  try {
-    printWallpaper();
-    printBackgroundColor();
-  } finally {
-    free(wallpaper.ptr);
-    CoUninitialize();
-  }
+  printWallpaper();
+  printBackgroundColor();
 }